Improve camera movements (mouse support), fixed point arithmetic in the inner loop

+++ src/stats.rs
@@ -4,11 +4,11 @@ use std::fmt::{Display, Formatter};
 #[derive(Default)]
 pub struct Stats {
     num_events: u32,
-    total_time: usize,
+    total_time: u32,
 }
 
 impl Stats {
-    pub fn add(self: &mut Self, time: usize) {
+    pub fn add(self: &mut Self, time: u32) {
         self.num_events += 1;
         self.total_time += time;
     }
